requirements.txt
discord.py>=1.3.3
beatifulsoup4==4.9.1
requests==2.24.0
